Separated all components and updated page
